backend/services/review-service/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.1.3
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
h11==0.16.0
httpx==0.28.1
idna==3.11
langchain==0.2.16
langchain-core==0.2.38
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.0
pymongo==4.10.1
python-multipart==0.0.20
python-dotenv==1.2.2
SQLAlchemy==2.0.48
starlette==0.52.1
tavily-python==0.7.12
typing_extensions==4.15.0
uvicorn==0.41.0
pymysql==1.1.0
kafka-python==2.0.2
